Add django debug toolbar
This adds the ability to run the app with a debug toolbar enabled.
It is disabled by default, and only enabled currently in docker compose
for development.

It can be enabled with a env variable, so if did want to enable it temporarily
in production for debugging we could do that.